Tagging X10 2.4 branch revision 27029 as SF_RELEASE_2_4_0_7 release of X10

+++ x10.runtime/src-java/x10/core/io/InputStream.java
@@ -117,7 +117,7 @@ public class InputStream extends Ref {
         }
     }
 
-    public void skip(long n) {
+    public void skip(int n) {
         try {
             stream.skip(n);
         } catch (java.io.IOException e) {
